requirements_agent.txt
langgraph>=1.0.5,<2
langchain-anthropic>=0.3,<1
langchain-google-genai>=2,<3
langchain-openai>=0.3,<1
PyYAML>=6,<7
python-dotenv>=1.0,<2
pytest>=8,<9
